feat: 좋아요 api를 클라이언트 용 | 서버 용으로 분리
