+++ src/fetcher/index.ts
@@ -1,7 +1,7 @@
 import { notFound } from 'next/navigation';
 import { backendFetch } from '@/fetcher/backendFetch.ts';
 import type { DefaultRes, SearchWord } from './types.ts';
-import { BackendFetchRes, ErrorRes, LoginData } from './types.ts';
+import { BackendFetchRes, LoginData } from './types.ts';
 
 export const getWordSearch = async (wordName: string) => {
   try {
@@ -37,40 +37,32 @@ export function isServer() {
   return typeof window === 'undefined';
 }
 
-export async function responseInterceptor(response: Response) {
-  if (isServer()) {
-    console.log('********* after sending request in server *********\n');
-    console.log(response.status);
-  }
-
-  const data = await response.json();
-
-  if (!response.ok) {
-    // NOTE: 서버사이드에서 호출 시에는 로깅합니다.
-    if (isServer()) {
-      if (400 <= response.status && response.status < 500) {
-        console.log('ClientError: ', response.status);
-      } else {
-        console.error('ServerError : ', response.status);
-      }
+export const updateLike = async (wordId: string) => {
+  try {
+    await backendFetch<BackendFetchRes<DefaultRes<never>>>(`/like/${wordId}`, {
+      method: 'PATCH',
+    });
+  } catch (e) {
+    if (e instanceof Error) {
+      console.log(e.message);
     }
-
-    //NOTE: 에러 시 내려오는 객체의 형식으로 type assertion 합니다.
-    const res = data as ErrorRes;
-
-    // NOTE: axios 기본 동작과 동일하게, response.ok가 아니면 Error 를 throw 합니다.
-    throw new Error(
-      JSON.stringify({
-        statusCode: response.status,
-        message: res.message,
-      }),
-    );
+    // NOTE: 발생할 수 있는 에러
+    // 401 => 권한 없음 => 로그인 모달
+    // 500 => 서버 에러
   }
+};
 
-  return {
-    status: response.status,
-    statusText: response.statusText,
-    headers: response.headers,
-    data,
-  };
-}
+export const deleteLike = async (wordId: string) => {
+  try {
+    await backendFetch<BackendFetchRes<DefaultRes<never>>>(`/like/${wordId}`, {
+      method: 'DELETE',
+    });
+  } catch (e) {
+    if (e instanceof Error) {
+      console.log(e.message);
+    }
+    // NOTE: 발생할 수 있는 에러
+    // 401 => 권한 없음 => 로그인 모달
+    // 500 => 서버 에러
+  }
+};
